stat-api/requirements.txt
cryptography==2.9.2
matplotlib==3.2.2
numpy==1.18.5
pandas==1.0.4
scikit-image==0.16.2
scikit-learn==0.24.1
scipy==1.5.0
pyjwt==1.7.1
apscheduler==3.6.3
fsspec
data-pipeline-api
elasticsearch
pymongo
dnspython
neo4j==4.1.1
pytest
marshmallow==3.9.0
pyamg
fastapi
uvicorn
pydantic
loguru
dependency-injector